+++ doc/whats-new.rst
@@ -53,6 +53,8 @@ Bug fixes
 - :py:func:`xarray.concat` can now concatenate variables present in some datasets but
   not others (:issue:`508`, :pull:`7400`).
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_ and `Scott Chamberlin <https://github.com/scottcha>`_.
+- Handle ``keep_attrs`` option in binary operators of :py:meth:`Dataset` (:issue:`7390`, :pull:`7391`).
+  By `Aron Gergely <https://github.com/arongergely>`_.
 
 Documentation
 ~~~~~~~~~~~~~